refactored the templates and package name
