+++ TaskManager/src/main/java/com/amalitechtaskmanager/exception/dummy.java
@@ -1,0 +1,4 @@
+package com.amalitechtaskmanager.exception;
+
+public class dummy {
+}
