+++ TaskManager/src/main/java/com/amalitechtaskmanager/handlers/auth/dummy.java
@@ -1,0 +1,4 @@
+package com.amalitechtaskmanager.handlers.auth;
+
+public class dummy {
+}
