+++ TaskManager/src/main/java/com/amalitechtaskmanager/handlers/task/dummy.java
@@ -1,0 +1,5 @@
+package com.amalitechtaskmanager.handlers.task;
+
+public class dummy {
+
+}
