+++ TaskManager/src/main/java/com/amalitechtaskmanager/model/dummy.java
@@ -1,0 +1,5 @@
+package com.amalitechtaskmanager.model;
+
+public class dummy {
+
+}
