+++ TaskManager/src/main/java/com/amalitechtaskmanager/utils/dummy.java
@@ -1,0 +1,4 @@
+package com.amalitechtaskmanager.utils;
+
+public class dummy {
+}
