--- TaskManager/src/main/java/taskmanager/exception/dummy.java
@@ -1,4 +1,0 @@
-package taskmanager.exception;
-
-public class dummy {
-}
