--- TaskManager/src/main/java/taskmanager/handlers/auth/dummy.java
@@ -1,4 +1,0 @@
-package taskmanager.handlers.auth;
-
-public class dummy {
-}
