--- TaskManager/src/main/java/taskmanager/handlers/task/dummy.java
@@ -1,4 +1,0 @@
-package taskmanager.handlers.task;
-
-public class dummy {
-}
