--- TaskManager/src/main/java/taskmanager/model/dummy.java
@@ -1,4 +1,0 @@
-package taskmanager.model;
-
-public class dummy {
-}
