--- TaskManager/src/main/java/taskmanager/utils/dummy.java
@@ -1,4 +1,0 @@
-package taskmanager.utils;
-
-public class dummy {
-}
