+++ TaskManager/src/test/java/com/amalitechtaskmanager/AppTest.java
@@ -1,4 +1,4 @@
-package taskmanager;
+package com.amalitechtaskmanager;
 
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyResponseEvent;
 import static org.junit.Assert.assertEquals;
