+++ sns-sqs-template.yaml
@@ -1,0 +1,93 @@
+AWSTemplateFormatVersion: '2010-09-09'
+Description: SNS and SQS Resources for Task Management System Backend
+
+Resources:
+  TaskAssignmentNotificationTopic:
+    Type: AWS::SNS::Topic
+    Properties:
+      TopicName: TaskAssignmentNotificationTopic.fifo
+      FifoTopic: true
+      ContentBasedDeduplication: true
+      Tags:
+        - Key: Component
+          Value: SNS-SQS
+
+  TaskDeadlineNotificationTopic:
+    Type: AWS::SNS::Topic
+    Properties:
+      TopicName: TaskDeadlineNotification
+      Tags:
+        - Key: Component
+          Value: SNS-SQS
+
+  ClosedTasksNotificationTopic:
+    Type: AWS::SNS::Topic
+    Properties:
+      TopicName: ClosedTasksNotification
+      Tags:
+        - Key: Component
+          Value: SNS-SQS
+
+  TaskCompleteNotificationTopic:
+    Type: AWS::SNS::Topic
+    Properties:
+      TopicName: TaskCompleteNotification
+      Tags:
+        - Key: Component
+          Value: SNS-SQS
+
+  EmailNotificationTopic:
+    Type: AWS::SNS::Topic
+    Properties:
+      TopicName: EmailNotificationTopic
+      DisplayName: "Task Notifications"
+      Tags:
+        - Key: Component
+          Value: SNS-SQS
+
+  TaskCreationDLQ:
+    Type: AWS::SQS::Queue
+    Properties:
+      QueueName: TaskCreationDLQ.fifo
+      FifoQueue: true
+      Tags:
+        - Key: Component
+          Value: SNS-SQS
+
+  TaskAssignmentQueue:
+    Type: AWS::SQS::Queue
+    Properties:
+      QueueName: TaskAssignmentQueue.fifo
+      FifoQueue: true
+      ContentBasedDeduplication: true
+      RedrivePolicy:
+        deadLetterTargetArn: !GetAtt TaskCreationDLQ.Arn
+        maxReceiveCount: 3
+      Tags:
+        - Key: Component
+          Value: SNS-SQS
+
+  TaskDeadlineQueue:
+    Type: AWS::SQS::Queue
+    Properties:
+      QueueName: TaskDeadlineQueue
+      Tags:
+        - Key: Component
+          Value: SNS-SQS
+
+Outputs:
+  TaskAssignmentNotificationTopicArn:
+    Description: SNS Task Assignment Notification Topic ARN
+    Value: !Ref TaskAssignmentNotificationTopic
+  TaskDeadlineNotificationTopicArn:
+    Description: SNS Task Deadline Notification Topic ARN
+    Value: !Ref TaskDeadlineNotificationTopic
+  ClosedTasksNotificationTopicArn:
+    Description: SNS Closed Tasks Notification Topic ARN
+    Value: !Ref ClosedTasksNotificationTopic
+  TaskCompleteNotificationTopicArn:
+    Description: SNS Task Complete Notification Topic ARN
+    Value: !Ref TaskCompleteNotificationTopic
+  TaskAssignmentQueueArn:
+    Description: SQS Task Assignment Queue ARN
+    Value: !GetAtt TaskAssignmentQueue.Arn
